+++ examples/blocking_async.rs
@@ -12,7 +12,7 @@ use bevy::{
     tasks::{AsyncComputeTaskPool, Task},
     DefaultPlugins,
 };
-//use bevy_prototype_debug_lines::{DebugLines, DebugLinesPlugin};
+use bevy_prototype_debug_lines::{DebugLines, DebugLinesPlugin};
 use bevy_rapier3d::prelude::{Collider, NoUserData, RapierConfiguration, RapierPhysicsPlugin};
 use futures_lite::future;
 use oxidized_navigation::{
@@ -26,7 +26,7 @@ fn main() {
         // Default Plugins
         .add_plugins(DefaultPlugins)
         // Debug Lines for drawing nav-mesh.
-        //.add_plugin(DebugLinesPlugin::default())
+        .add_plugin(DebugLinesPlugin::default())
         .insert_resource(NavMeshSettings {
             cell_width: 0.25,
             cell_height: 0.1,
@@ -52,12 +52,16 @@ fn main() {
             ..Default::default()
         })
         .insert_resource(AsyncPathfindingTasks::default())
-        .add_startup_system(setup_world_system)
-        .add_startup_system(info_system)
-        .add_system(run_blocking_pathfinding)
-        .add_system(run_async_pathfinding)
-        .add_system(poll_pathfinding_tasks_system)
-        //.add_system(draw_nav_mesh_system)
+        .add_startup_systems((
+            setup_world_system,
+            info_system
+        ))
+        .add_systems((
+            run_blocking_pathfinding,
+            run_async_pathfinding,
+            poll_pathfinding_tasks_system,
+            draw_nav_mesh_system
+        ))
         .run();
 }
 
@@ -71,7 +75,7 @@ fn run_blocking_pathfinding(
     keys: Res<Input<KeyCode>>,
     nav_mesh_settings: Res<NavMeshSettings>,
     nav_mesh: Res<NavMesh>,
-    //mut lines: ResMut<DebugLines>,
+    mut lines: ResMut<DebugLines>,
 ) {
     if !keys.just_pressed(KeyCode::B) {
         return;
@@ -98,7 +102,7 @@ fn run_blocking_pathfinding(
                 match perform_string_pulling_on_path(&nav_mesh, start_pos, end_pos, &path) {
                     Ok(string_path) => {
                         info!("String path (BLOCKING): {:?}", string_path);
-                        //draw_path(&string_path, &mut lines, Color::RED);
+                        draw_path(&string_path, &mut lines, Color::RED);
                     }
                     Err(error) => error!("Error with string path: {:?}", error),
                 };
@@ -154,13 +158,13 @@ fn run_async_pathfinding(
 // Poll existing tasks.
 fn poll_pathfinding_tasks_system(
     mut pathfinding_task: ResMut<AsyncPathfindingTasks>,
-    //mut lines: ResMut<DebugLines>,
+    mut lines: ResMut<DebugLines>,
 ) {
     // Go through and remove completed tasks.
     pathfinding_task.tasks.retain_mut(|task| {
         if let Some(string_path) = future::block_on(future::poll_once(task)).unwrap_or(None) {
             info!("Async path task finished with result: {:?}", string_path);
-            //draw_path(&string_path, &mut lines, Color::BLUE);
+            draw_path(&string_path, &mut lines, Color::BLUE);
 
             false
         } else {
@@ -210,11 +214,11 @@ async fn async_path_find(
 
 // General use path draw.
 
-/*fn draw_path(path: &[Vec3], lines: &mut DebugLines, color: Color) {
+fn draw_path(path: &[Vec3], lines: &mut DebugLines, color: Color) {
     for (a, b) in path.iter().zip(path.iter().skip(1)) {
         lines.line_colored(*a, *b, 15.0, color);
     }
-}*/
+}
 
 //
 //  Draw Nav-mesh.
@@ -222,7 +226,7 @@ async fn async_path_find(
 //
 //  This uses the crate ``bevy_prototype_debug_lines``.
 //
-/*fn draw_nav_mesh_system(
+fn draw_nav_mesh_system(
     keys: Res<Input<KeyCode>>,
     nav_mesh: Res<NavMesh>,
     mut lines: ResMut<DebugLines>,
@@ -256,7 +260,7 @@ async fn async_path_find(
             }
         }
     }
-}*/
+}
 
 fn setup_world_system(
     mut commands: Commands,
@@ -272,7 +276,7 @@ fn setup_world_system(
             ..default()
         },
         Collider::cuboid(25.0, 0.1, 25.0),
-        NavMeshAffector::default(), // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
+        NavMeshAffector, // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
     ));
 
     // Cube
@@ -284,7 +288,7 @@ fn setup_world_system(
             ..default()
         },
         Collider::cuboid(1.25, 1.25, 1.25),
-        NavMeshAffector::default(), // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
+        NavMeshAffector, // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
     ));
 
     // Thin wall
@@ -296,7 +300,7 @@ fn setup_world_system(
             ..default()
         },
         Collider::cuboid(0.05, 0.05, 0.05),
-        NavMeshAffector::default(), // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
+        NavMeshAffector, // Only entities with a NavMeshAffector component will contribute to the nav-mesh.
     ));
 
     // light